Fix all remaining Clang strict compiler warnings (25/25 files clean)
Achieved 100% warning-free compilation on both GCC and Clang with strictest flags:
- GCC: 25/25 files ✓
- Clang: 25/25 files ✓
- All flags: -Wall -Wextra -Wpedantic -Wsign-conversion -Wconversion -Werror

Files Fixed:
============

**Header Files (Type Casting):**
- src/varintSplitFull.h: Added (uint8_t) casts in varintSplitFullLengthVAR_, varintSplitFullPut_, varintSplitFullGet_, varintSplitFullReversedPutReversed_, varintSplitFullReversedPutForward_, varintSplitFullReversedGet_ macros
- src/varintSplitFull16.h: Added (uint8_t) casts in varintSplitFull16LengthVAR_, varintSplitFull16Put_, and (uint64_t) casts in varintSplitFull16Get_ macros
- src/varintSplit.h: Added (uint8_t) casts in varintSplitLengthVAR_, varintSplitPut_, varintSplitReversedPutReversed_, varintSplitReversedPutForward_, and (uint64_t) casts in varintSplitGet_, varintSplitReversedGet_ macros
- src/varintSplitFullNoZero.h: Added (uint8_t) casts in varintSplitFullNoZeroReversedPutReversed_, varintSplitFullNoZeroReversedPutForward_, and (uint64_t) casts in varintSplitFullNoZeroReversedGet_ macros
- src/varintPackedTest.c: Added (uint32_t) and (uint16_t) casts for all varintPacked function calls throughout all test loops

**Type Casting Patterns Applied:**
1. **Put Macros:** All dst[] byte assignments wrapped with (uint8_t) casts to prevent truncation warnings
2. **Get Macros:** All bit shift operations wrapped with (uint64_t) casts to prevent sign conversion warnings
3. **LengthVAR Macros:** All encodedLen assignments wrapped with (uint8_t) casts
4. **Test Files:** All function parameters cast to correct types (uint32_t for indices, uint16_t for values, uint64_t for performance counters)

**Test Verification:**
- All 8 tests pass without sanitizers
- All 8 tests pass with AddressSanitizer
- All 8 tests pass with UndefinedBehaviorSanitizer
- All 8 tests pass with BOTH sanitizers combined

This completes the comprehensive compiler warning elimination project across
all varint encoding implementations.

+++ src/varintCompare.c
@@ -99,7 +99,7 @@ int32_t main(int argc, char **argv) {
         for (i = 0; i < maxLoop; i++) {
             GIVE_XZ;
 
-            int n1 = varintTaggedPut64(z, x);
+            int n1 = (int)varintTaggedPut64(z, x);
             uint64_t y = 0;
             varintTaggedGet(z, n1, &y);
             assert(x == y);
@@ -146,10 +146,10 @@ int32_t main(int argc, char **argv) {
         for (i = 0; i < maxLoop; i++) {
             GIVE_XZ;
 
-            int32_t n1 = varintChainedPutVarint(z, x);
+            int32_t n1 = (int32_t)varintChainedPutVarint(z, x);
             assert(n1 >= 1 && n1 <= 9);
             uint64_t y = 0;
-            int32_t n2 = varintChainedGetVarint(z, &y);
+            int32_t n2 = (int32_t)varintChainedGetVarint(z, &y);
             assert(n1 == n2);
             assert(x == y);
 
@@ -181,10 +181,10 @@ int32_t main(int argc, char **argv) {
         for (i = 0; i < maxLoop; i++) {
             GIVE_XZ;
 
-            int32_t n1 = varintChainedSimpleEncode64(z, x);
+            int32_t n1 = (int32_t)varintChainedSimpleEncode64(z, x);
             assert(n1 >= 1 && n1 <= 9);
             uint64_t y = 0;
-            int32_t n2 = varintChainedSimpleDecode64(z, &y);
+            int32_t n2 = (int32_t)varintChainedSimpleDecode64(z, &y);
             assert(n1 == n2);
             assert(x == y);
 
@@ -200,10 +200,10 @@ int32_t main(int argc, char **argv) {
         for (i = 0; i < maxLoop; i++) {
             GIVE_SMALL_XZ
 
-            int32_t n1 = varintChainedSimpleEncode32(z, (uint32_t)x);
+            int32_t n1 = (int32_t)varintChainedSimpleEncode32(z, (uint32_t)x);
             assert(n1 >= 1 && n1 <= 5);
             uint32_t y = 0;
-            int32_t n2 = varintChainedSimpleDecode32(z, &y);
+            int32_t n2 = (int32_t)varintChainedSimpleDecode32(z, &y);
             assert(n1 == n2);
             assert(x == (uint64_t)y);
